requirements.txt
hikari
hikari-crescent
hikari-miru
python-dotenv
motor
lavalink_rs